+++ src/info.py
@@ -2,5 +2,5 @@ class Info:
     Author = 'Andreas Pfau'
     Domain = 'apfau.de'
     AppName = 'S-Parameter Viewer'
-    AppVersionStr = "0.30b1" # see https://peps.python.org/pep-0440/
-    AppDateStr = "2025-04-14"
+    AppVersionStr = "0.31b1"  # see https://peps.python.org/pep-0440/
+    AppDateStr = "2025-05-09"